Refactor code to follow a naming convention
