+++ test/fuzzing/client/client-fuzz-body.js
