+++ test/fuzzing/client/client-fuzz-options.js
